requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pandas>=2.0.0
requests>=2.31.0

# Optional - Database drivers
# psycopg2-binary>=2.9.9  # PostgreSQL
# mysql-connector-python>=8.3.0  # MySQL
